pipeline/requirements.txt
requests==2.32.3
pandas==2.2.3
scikit-learn==1.5.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
